statistic service added

+++ src/main/java/com/getir/casestudy/service/BookService.java
@@ -1,11 +1,11 @@
 package com.getir.casestudy.service;
 
 import com.getir.casestudy.domain.Book;
+import com.getir.casestudy.exception.BookNotFoundException;
 import com.getir.casestudy.model.request.BookCreateRequest;
 import com.getir.casestudy.model.request.BookStockUpdateRequest;
 import com.getir.casestudy.model.response.BookResponse;
 import com.getir.casestudy.repository.IBookRepository;
-import lombok.AllArgsConstructor;
 import lombok.NoArgsConstructor;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -14,21 +14,22 @@ import org.springframework.stereotype.Service;
 import javax.persistence.EntityNotFoundException;
 
 @Service
-@AllArgsConstructor
-@NoArgsConstructor
 public class BookService {
     private final Logger log = LoggerFactory.getLogger(BookService.class);
 
-    private IBookRepository bookRepository;
+    private final IBookRepository bookRepository;
+
+    public BookService(IBookRepository bookRepository) {
+        this.bookRepository = bookRepository;
+    }
 
     public BookResponse createBook(BookCreateRequest request) {
         log.debug("BookService - createBook started");
-        Book book = new Book();
-        book.setTitle(request.getTitle());
-        book.setDescription(request.getDescription());
-        book.setAuthor(request.getAuthor());
-        book.setPrice(request.getPrice());
-        book.setRemainingStock(request.getRemainingStock());
+        Book book = Book.builder().title(request.getTitle())
+                .description(request.getDescription())
+                .author(request.getAuthor())
+                .price(request.getPrice())
+                .remainingStock(request.getRemainingStock()).build();
 
         bookRepository.save(book);
 
@@ -41,7 +42,7 @@ public class BookService {
 
     public BookResponse getBook(String id) {
         log.debug("BookService - getBook started");
-        Book book = bookRepository.findById(id).orElseThrow(() -> new EntityNotFoundException(String.valueOf(id)));
+        Book book = bookRepository.findById(id).orElseThrow(() -> new EntityNotFoundException(id));
         BookResponse response = new BookResponse();
         response.setBook(book);
 
@@ -51,8 +52,9 @@ public class BookService {
 
     public BookResponse updateBookStock(BookStockUpdateRequest request) {
         log.debug("BookService - updateBookStock started");
-        Book book = bookRepository.findById(request.getId()).orElseThrow(() -> new EntityNotFoundException(String.valueOf(request.getId())));
+        Book book = bookRepository.findBookById(request.getId());
 
+        if (book == null) { throw new BookNotFoundException();}
         book.setRemainingStock(request.getRemainingStock());
 
         bookRepository.save(book);
@@ -64,15 +66,14 @@ public class BookService {
         return response;
     }
 
-    public BookResponse deleteBook(String id) {
+    public String deleteBook(String id) {
         log.debug("BookService - deleteBook started");
-        BookResponse response = new BookResponse();
-        Book book = bookRepository.findById(id).orElseThrow(() -> new EntityNotFoundException(String.valueOf(id)));
+        Book book = bookRepository.findBookById(id);
+        if (book == null) { throw new BookNotFoundException();}
 
-        response.setBook(book);
         bookRepository.delete(book);
 
         log.debug("BookService - deleteBook - book is deleted");
-        return response;
+        return "Book is deleted";
     }
 }
